新シーン追加・エディタUI改善・DirectXTexビルド更新
DemoGameScene2.sceneの追加、エディタUI/UXの向上（ヒエラルキーのポップアップメニュー導入等）、カメラやシェーダの修正、imguiレイアウト調整、DirectXTex.libのビルド成果物・シンボル情報の追加・更新を行いました。これにより、ゲームステージ構成の拡充、エディタ操作性の向上、描画処理の正確化、ライブラリ機能の強化が図られています。

+++ InvaderGame/GameSystem.cpp
@@ -130,7 +130,7 @@ void GameSystem::Execute()
 	cd_rendering->categories["CamSetting"] = subCD_camSetting;
 	TimeMeasurer::Start();
 	DirectX::XMVECTOR camPos;
-	int camSize = 1;
+	float camSize = 1;
 	if (SceneEditor::GetIsEditMode())
 	{
 		camPos = EditorCamera::GetPosition().ToXMVECTOR();
